Move to making upthread_lib_init a constructor
The idea being that if you link this in at all, you are probably
planning on using threads.  If we don't do this, then the main thread
changes identiity (via a call to pthread_self()) after the first upthread
is created (because it gets upgraded to a upthread itself at that
point).  We want it to always be a upthread.

Also fixed a bug related to the new uthread_lib_init() interface.
It returns void now, so we can't assert it returning 0.