Stripe to avoid TLS conflicts
Instead of just walking through all of the cores and inserting threads in the
queues without any sort of intelligence, now walk through and try and place
them in queues where they wont conflict in the L1 cache with the tls of their
vcores. The algorithm makes sure that the threads are load balanced in the
process.  t lDoing so allows us to remove the extra paramter from the
set_num_vcores() call since we decide where best to put the threads.