Separate strategies for striping vcore queues
Previously the default was to always strip the vcore queues in a tls-aware
fashion to reduce contention.  Under certain circumstances, this can actually
harm things because it moves data farm from where it is being processed by
other threads.  This commit reintroduces the basic dumb way of striping as
a spearate call you can make.