Set prefered vcq to current vcq, not previous
This was a bug that resulted in a thread never actually migrating when pulled
from another queue.  It would just reset its preferred vcq back to the one it
was pulled from, instead of updating to the new one.  This meant it was placed
back in the old queue once it yielded, rather than the new one.  Now it is
migrated correctly.