Need to link this way to get sched_ops correct
The sched_ops variable is weak in -lparlib, and overrriden in lithe to
be strong.  However, the app itself needs to link in -lparlib because it
calls vcore_id() which comes from parlib.  Since the application itself
links in parlib, its weak symbol takes precendance over the strong
symbol in -lithe (which is stupid in my opinion).

Even if we have -lithe in the application, but leave -lithe as part of
the LDADD, we still have problems. Only when we *omit* -lithe from
LDADD and *include* it in the app does it link and run as desired.

This is obviously not how we want it in the long run.  Maybe someone
else has a better idea of how to set this up. Maybe it's time we move
away from eak symbols.