Path to actual upthread library must be relative.
If the library is installed in the non-default location, it will pick up
the wrong definitions.  I saw this manifest as a bug with the
pthread_t struct being interpreted as a normal upthread_mutex_t struct
from the pvcq library rather than a lithe_mutex_t from lithe.  It was
all becuase the upthread.h it was picking up was the wrong one...