Make definitions static!
The non-static versions were interfering (silently) with the list of the
same name in lithe/fork_join_sched.c. The bug manifested as an
incorrectly sized upthread_lithe_context_t struct. Specifically, the
offset was set to 512, when the size of a upthread_lithe_context_t
struct is 640 bytes!  It's because it was pulling a
lithe_fork_join_context_t out of the list instead and trying to use it
as a upthread_lithe_context_t.  All because of a naming collision.  Fix
is easy - make it static (in all places).