Add UPTHREAD_SPIN_COUNT environment variable
This count allows you to adjust how many times sched_entry() will spin
befor eactually yielding a vcore back to the system.  The value itself
is stored in an (int), so setting it to -1 will (essentially) give you
an infinite count, menaing that a vcore will never be yielded back.
This count is useful for allowing the upthread library to request vcores
itself, but only yield them back *very* rarely (as opposed to using the
upthread_can_vcore_request() call, which mieans you never request or
yeild within the library itself).  This is a nice compromise between
letting hte library explciitly request, but not yield back to
frequently.

In the future we can consider making this value settable via an extended
API call rather than an environment variable.